Refactor Elasticsearch backend classes to drop version 2
