+++ wagtail/search/tests/test_elasticsearch5_backend.py
@@ -743,7 +743,7 @@ class TestElasticsearch5MappingInheritance(TestCase):
         self.assertDictEqual(document, expected_result)
 
 
-@mock.patch('wagtail.search.backends.elasticsearch2.Elasticsearch')
+@mock.patch('wagtail.search.backends.elasticsearch5.Elasticsearch')
 class TestBackendConfiguration(TestCase):
     def test_default_settings(self, Elasticsearch):
         Elasticsearch5SearchBackend(params={})
