+++ wagtail/search/tests/test_elasticsearch7_backend.py
@@ -741,7 +741,7 @@ class TestElasticsearch7MappingInheritance(TestCase):
         self.assertDictEqual(document, expected_result)
 
 
-@mock.patch('wagtail.search.backends.elasticsearch2.Elasticsearch')
+@mock.patch('wagtail.search.backends.elasticsearch5.Elasticsearch')
 class TestBackendConfiguration(TestCase):
     def test_default_settings(self, Elasticsearch):
         Elasticsearch7SearchBackend(params={})
